Added the recursive option and the possibility to select a target directory
The recursive option is added.
The optional argument target is also added. The default behavior is to clean the current directory, but if one or more targets are specified, they are cleaned instead of the current directory.
Some function names have changed to be more specific.
The list_files_full_path function (now list_files) have been remake to integrate the ability to make a recursion.
This is now the v0.3