Changed a function to make it more generic
The ask_user_before_remove is now replaced by question
It's not a file that is asked as parameter, but a string representing the question
the user must answer "y" or "n" at the question, and the function loop while the answer is not "y" or "n".
Return True if the the answer is "y".