Added an option to change the pattern file
The (-c / --configuration) option allow to choose one or more file in which the programm will read the patterns to use.
These files must be in $HOME/.config/clean/ directory.

Fixed a bug with the patterns. Those including the begginning, such as "#*#" wasn't working, but it's fixed.

Modified the select_files function to accept a list of files.

Modified README.md to include informations about the new option.